+++ CHANGELOG.md
@@ -4,6 +4,19 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+v0.10.0
+------------------------------
+*February 6, 2018*
+
+### Added
+- Allow user to specify location of error message per field
+
+### Changed
+- Only validate fields on blur/keyup if they are not empty
+- dateInFuture: Only validating fields when they have both been interacted with
+- dateInFuture: Allowing current year/month to be valid
+
+
 v0.9.0
 ------------------------------
 *January 30, 2018*